refactor: 패키지 구조를 모델 단위로 변경
특정 객체는 어떠한 객체의 하위로 판단
* Product
  * Category
* User
  * Role(UserRole)
* Wish
  * User의 하위로 둘지 고민
