--- src/main/java/gift/controller/UserController.java
@@ -1,34 +1,0 @@
-package gift.controller;
-
-import gift.dto.user.request.UserLoginRequest;
-import gift.dto.user.request.UserRegisterRequest;
-import gift.dto.user.response.UserResponse;
-import gift.service.UserService;
-import jakarta.validation.Valid;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/api/users")
-public class UserController {
-
-    private final UserService userService;
-
-    public UserController(UserService userService) {
-        this.userService = userService;
-    }
-
-    @PostMapping("register")
-    public ResponseEntity<UserResponse> register(@RequestBody @Valid UserRegisterRequest request) {
-        return ResponseEntity.ok(userService.registerUser(request));
-    }
-
-    @PostMapping("login")
-    public ResponseEntity<UserResponse> login(@RequestBody @Valid UserLoginRequest userRequest) {
-        return ResponseEntity.ok(userService.loginUser(userRequest));
-    }
-
-}
