--- src/main/java/gift/controller/viewcontroller/MainViewController.java
@@ -1,13 +1,0 @@
-package gift.controller.viewcontroller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class MainViewController {
-
-    @GetMapping
-    public String index() {
-        return "index";
-    }
-}
