--- src/main/java/gift/controller/viewcontroller/UserViewController.java
@@ -1,21 +1,0 @@
-package gift.controller.viewcontroller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@Controller
-@RequestMapping("/users")
-public class UserViewController {
-
-    @GetMapping("/register")
-    public String register() {
-        return "register_form";
-    }
-
-    @GetMapping("/login")
-    public String login() {
-        return "login_form";
-    }
-
-}
