--- src/main/java/gift/dto/product/response/ProductResponse.java
@@ -1,11 +1,0 @@
-package gift.dto.product.response;
-
-public record ProductResponse(
-    Long id,
-    String name,
-    Integer price,
-    String imageUrl,
-    String categoryName
-) {
-
-}
