--- src/main/java/gift/dto/user/request/UserLoginRequest.java
@@ -1,11 +1,0 @@
-package gift.dto.user.request;
-
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotEmpty;
-
-public record UserLoginRequest(
-    @Email String email,
-    @NotEmpty String password
-) {
-
-}
