--- src/main/java/gift/dto/user/response/UserResponse.java
@@ -1,7 +1,0 @@
-package gift.dto.user.response;
-
-public record UserResponse(
-    String token
-) {
-
-}
