--- src/main/java/gift/dto/wish/response/WishResponse.java
@@ -1,11 +1,0 @@
-package gift.dto.wish.response;
-
-import gift.entity.Product;
-
-public record WishResponse(
-    Long id,
-    Product product,
-    Integer quantity
-) {
-
-}
