--- src/main/java/gift/entity/Product.java
@@ -1,124 +1,0 @@
-package gift.entity;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.ForeignKey;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
-
-@Entity
-@Table(name = "products")
-public class Product {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(length = 15, nullable = false)
-    private String name;
-
-    @Column(nullable = false)
-    private Integer price;
-
-    @Column(nullable = false)
-    private String imageUrl;
-
-    @ManyToOne
-    @JoinColumn(name = "category_id", nullable = false,
-        foreignKey = @ForeignKey(name = "fk_products_category_id_ref_categories_id"))
-    private Category category;
-
-    protected Product() {
-    }
-
-    private Product(Builder builder) {
-        this.id = builder.id;
-        this.name = builder.name;
-        this.price = builder.price;
-        this.imageUrl = builder.imageUrl;
-        this.category = builder.category;
-    }
-
-    public static Builder builder() {
-        return new Builder();
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public Integer getPrice() {
-        return price;
-    }
-
-    public String getImageUrl() {
-        return imageUrl;
-    }
-
-    public String getCategoryName() {
-        return category.getName();
-    }
-
-    public void changeName(String name) {
-        this.name = name;
-    }
-
-    public void changePrice(Integer price) {
-        this.price = price;
-    }
-
-    public void changeImageUrl(String imageUrl) {
-        this.imageUrl = imageUrl;
-    }
-
-    public void changeCategory(Category category) {
-        this.category = category;
-    }
-
-    public static class Builder {
-
-        private Long id;
-        private String name;
-        private Integer price;
-        private String imageUrl;
-        private Category category;
-
-        public Builder id(Long id) {
-            this.id = id;
-            return this;
-        }
-
-        public Builder name(String name) {
-            this.name = name;
-            return this;
-        }
-
-        public Builder price(Integer price) {
-            this.price = price;
-            return this;
-        }
-
-        public Builder imageUrl(String imageUrl) {
-            this.imageUrl = imageUrl;
-            return this;
-        }
-
-        public Builder category(Category category) {
-            this.category = category;
-            return this;
-        }
-
-        public Product build() {
-            return new Product(this);
-        }
-    }
-
-}
