--- src/main/java/gift/entity/User.java
@@ -1,105 +1,0 @@
-package gift.entity;
-
-import jakarta.persistence.CascadeType;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.OneToMany;
-import jakarta.persistence.Table;
-import jakarta.persistence.UniqueConstraint;
-import java.util.Set;
-
-@Entity
-@Table(
-    name = "users",
-    uniqueConstraints = @UniqueConstraint(columnNames = {"email"}, name = "uk_users")
-)
-public class User {
-
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false)
-    private String email;
-
-    @Column(nullable = false)
-    private String password;
-
-    @OneToMany(mappedBy = "user", fetch = FetchType.LAZY, cascade = CascadeType.ALL, orphanRemoval = true)
-    private Set<UserRole> userRoles;
-
-    @OneToMany(mappedBy = "user", cascade = CascadeType.REMOVE, orphanRemoval = true)
-    private Set<Wish> wishes;
-
-    protected User() {
-    }
-
-    private User(Builder builder) {
-        this.id = builder.id;
-        this.email = builder.email;
-        this.password = builder.password;
-        this.wishes = builder.wishes;
-    }
-
-    public static Builder builder() {
-        return new Builder();
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public Set<UserRole> getRoles() {
-        return userRoles;
-    }
-
-    public void changePassword(String password) {
-        this.password = password;
-    }
-
-    public Set<Wish> wishes() {
-        return wishes;
-    }
-
-    public static class Builder {
-
-        private Long id;
-        private String email;
-        private String password;
-        private Set<Wish> wishes;
-
-        public Builder id(Long id) {
-            this.id = id;
-            return this;
-        }
-
-        public Builder email(String email) {
-            this.email = email;
-            return this;
-        }
-
-        public Builder password(String password) {
-            this.password = password;
-            return this;
-        }
-
-        public Builder wishes(Set<Wish> wishes) {
-            this.wishes = wishes;
-            return this;
-        }
-
-        public User build() {
-            return new User(this);
-        }
-    }
-
-}
