--- src/main/java/gift/entity/UserRole.java
@@ -1,33 +1,0 @@
-package gift.entity;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.ForeignKey;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
-
-@Entity
-@Table(name = "user_role")
-public class UserRole {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne
-    @JoinColumn(name = "user_id", nullable = false,
-        foreignKey = @ForeignKey(name = "fk_user_role_user_id_ref_users_id"))
-    private User user;
-
-    @ManyToOne
-    @JoinColumn(name = "role_id", nullable = false,
-        foreignKey = @ForeignKey(name = "fk_user_role_role_id_ref_roles_id"))
-    private Role role;
-
-    public Role getRole() {
-        return role;
-    }
-}
