+++ src/main/java/gift/product/dto/category/request/CreateCategoryRequest.java
@@ -1,11 +1,11 @@
-package gift.dto.category.request;
+package gift.product.dto.category.request;
 
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
 
-public record UpdateCategoryRequest(
-    @NotEmpty
+public record CreateCategoryRequest(
     @NotNull
+    @NotEmpty
     String name
 ) {
