+++ src/main/java/gift/product/dto/category/request/UpdateCategoryRequest.java
@@ -1,11 +1,11 @@
-package gift.dto.category.request;
+package gift.product.dto.category.request;
 
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
 
-public record CreateCategoryRequest(
-    @NotNull
+public record UpdateCategoryRequest(
     @NotEmpty
+    @NotNull
     String name
 ) {
