+++ src/main/java/gift/product/dto/category/response/CategoryResponse.java
@@ -1,4 +1,4 @@
-package gift.dto.category.response;
+package gift.product.dto.category.response;
 
 public record CategoryResponse(
     Long id,
