+++ src/main/java/gift/product/entity/Category.java
@@ -1,4 +1,4 @@
-package gift.entity;
+package gift.product.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
