--- src/main/java/gift/repository/ProductRepository.java
@@ -1,10 +1,0 @@
-package gift.repository;
-
-import gift.entity.Product;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ProductRepository extends JpaRepository<Product, Long> {
-
-}
