--- src/main/java/gift/repository/WishRepository.java
@@ -1,16 +1,0 @@
-package gift.repository;
-
-import gift.entity.Wish;
-import java.util.List;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface WishRepository extends JpaRepository<Wish, Long> {
-
-    Page<Wish> findByUserId(Long userId, Pageable pageable);
-
-    List<Wish> findByUserId(Long userId);
-}
